+++ simple_history/tests/tests/test_models.py
@@ -1,17 +1,9 @@
 from __future__ import unicode_literals
 
 from datetime import datetime, timedelta
-try:
-    from unittest import skipUnless
-except ImportError:
-    from unittest2 import skipUnless
+import unittest
 
 import django
-try:
-    from django.contrib.auth import get_user_model
-    User = get_user_model()
-except ImportError:  # django 1.4 compatibility
-    from django.contrib.auth.models import User
 from django.db import models
 from django.db.models.loading import get_model
 from django.db.models.fields.proxy import OrderWrt
@@ -25,9 +17,21 @@ from ..models import (
     Person, FileModel, Document, Record, Book, HistoricalPoll, Library, State,
     AbstractBase, ConcreteAttr, ConcreteUtil, SelfFK, Temperature, WaterLevel,
     ExternalModel1, ExternalModel3, UnicodeVerboseName, HistoricalChoice,
-    HistoricalState, HistoricalCustomFKError, Series, SeriesWork, PollInfo
+    HistoricalState, HistoricalCustomFKError, Series, SeriesWork, PollInfo,
+    UserAccessorDefault, UserAccessorOverride, Employee, Country, Province,
+    City
 )
 from ..external.models import ExternalModel2, ExternalModel4
+
+try:
+    from unittest import skipUnless
+except ImportError:
+    from unittest2 import skipUnless
+try:
+    from django.contrib.auth import get_user_model
+    User = get_user_model()
+except ImportError:  # django 1.4 compatibility
+    from django.contrib.auth.models import User
 
 today = datetime(2021, 1, 1, 10, 0)
 tomorrow = today + timedelta(days=1)
@@ -254,6 +258,21 @@ class HistoricalRecordsTest(TestCase):
         self.assertEqual([m.fk_id for m in model.history.all()],
                          [other.id, model.id, None])
 
+    def test_to_field_foreign_key_save(self):
+        country = Country.objects.create(code='US')
+        country2 = Country.objects.create(code='CA')
+        province = Province.objects.create(country=country)
+        province.country = country2
+        province.save()
+        self.assertEqual([c.country_id for c in province.history.all()],
+                         [country2.code, country.code])
+
+    def test_db_column_foreign_key_save(self):
+        country = Country.objects.create(code='US')
+        city = City.objects.create(country=country)
+        country_field = City._meta.get_field('country')
+        self.assertTrue(getattr(country_field, 'db_column') in str(city.history.all().query))
+
     def test_raw_save(self):
         document = Document()
         document.save_base(raw=True)
@@ -330,7 +349,7 @@ class RegisterTest(TestCase):
         self.assertEqual(expected, str(voter.history.all()[0])[:len(expected)])
 
 
-class CreateHistoryModelTests(TestCase):
+class CreateHistoryModelTests(unittest.TestCase):
 
     def test_create_history_model_with_one_to_one_field_to_integer_field(self):
         records = HistoricalRecords()
@@ -493,12 +512,20 @@ class HistoryManagerTest(TestCase):
             self.assertRaises(TypeError, HistoricalRecords, bases=bases)
 
     def test_import_related(self):
-        field_object = HistoricalChoice._meta.get_field_by_name('poll_id')[0]
-        self.assertEqual(field_object.related.model, Choice)
+        field_object = HistoricalChoice._meta.get_field_by_name('poll')[0]
+        try:
+            related_model = field_object.rel.related_model
+        except AttributeError:  # Django<1.8
+            related_model = field_object.related.model
+        self.assertEqual(related_model, HistoricalChoice)
 
     def test_string_related(self):
-        field_object = HistoricalState._meta.get_field_by_name('library_id')[0]
-        self.assertEqual(field_object.related.model, State)
+        field_object = HistoricalState._meta.get_field_by_name('library')[0]
+        try:
+            related_model = field_object.rel.related_model
+        except AttributeError:  # Django<1.8
+            related_model = field_object.related.model
+        self.assertEqual(related_model, HistoricalState)
 
     @skipUnless(django.get_version() >= "1.7", "Requires 1.7 migrations")
     def test_state_serialization_of_customfk(self):
@@ -658,7 +685,8 @@ class TestLatest(TestCase):
     """"Test behavior of `latest()` without any field parameters"""
 
     def setUp(self):
-        poll = Poll.objects.create(question="Does `latest()` work?", pub_date=yesterday)
+        poll = Poll.objects.create(
+            question="Does `latest()` work?", pub_date=yesterday)
         poll.pub_date = today
         poll.save()
 
@@ -689,3 +717,37 @@ class TestLatest(TestCase):
             {'pk': 2, 'history_date': yesterday},
         ])
         assert HistoricalPoll.objects.latest().pk == 1
+
+
+class TestUserAccessor(unittest.TestCase):
+
+    def test_accessor_default(self):
+        register(UserAccessorDefault)
+        assert not hasattr(User, 'historicaluseraccessordefault_set')
+
+    def test_accessor_override(self):
+        register(UserAccessorOverride, user_related_name='my_history_model_accessor')
+        assert hasattr(User, 'my_history_model_accessor')
+
+
+class TestMissingOneToOne(TestCase):
+
+    def setUp(self):
+        self.manager1 = Employee.objects.create()
+        self.manager2 = Employee.objects.create()
+        self.employee = Employee.objects.create(manager=self.manager1)
+        self.employee.manager = self.manager2
+        self.employee.save()
+        self.manager1.delete()
+
+    def test_history_is_complete(self):
+        historical_manager_ids = list(self.employee.history.order_by('pk')
+                                      .values_list('manager_id', flat=True))
+        self.assertEqual(historical_manager_ids, [1, 2])
+
+    def test_restore_employee(self):
+        historical = self.employee.history.order_by('pk')[0]
+        original = historical.instance
+        self.assertEqual(original.manager_id, 1)
+        with self.assertRaises(Employee.DoesNotExist):
+            original.manager